+++ .gitignore
@@ -21,6 +21,7 @@ out/
 
 
 # Cruft
+*.hprof
 .DS_Store
 .vscode
 epub
@@ -29,3 +30,4 @@ book.epub
 # deploy secrets and local configuration
 secring.gpg
 local.properties
+localRepo
