+++ gradle.properties
@@ -1,6 +1,6 @@
-# keep below 2 until ready to release v2.0 properly
 group=com.jillesvangurp
+# keep below 2 until ready to release v2.0 properly
 libraryVersion=1.99.0
 kotlin.mpp.stability.nowarn=true
-
-# set in local.properties or override with -P
+# dokka needs this
+org.gradle.jvmargs=-XX:MaxMetaspaceSize=512m
